refactor: useAuth에서 에러 처리를 하도록 변경
