+++ src/constants/validation.ts
@@ -1,0 +1,3 @@
+export const LOGIN = {
+  MIN_PASSWORD_LENGTH: 8,
+};
